Fixups to the action tutorials.
