--- source/Tutorials/Actions/client_0.c++
@@ -1,50 +1,0 @@
-#include <memory>
-#include <string>
-#include <sstream>
-
-#include "action_tutorials_interfaces/action/fibonacci.hpp"
-
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_action/rclcpp_action.hpp"
-#include "rclcpp_components/register_node_macro.hpp"
-
-namespace action_tutorials_cpp
-{
-class FibonacciActionClient : public rclcpp::Node
-{
-public:
-  explicit FibonacciActionClient(const rclcpp::NodeOptions & options)
-  : Node("fibonacci_action_client", options)
-  {
-
-  }
-
-  void send_goal()
-  {
-    
-  }
-private:
-  rclcpp_action::Client<Fibonacci>::SharedPtr client_ptr_;
-  rclcpp::TimerBase::SharedPtr timer_;
-
-  void goal_response_callback(std::shared_future<GoalHandleFibonacci::SharedPtr> future)
-  {
-
-  }
-
-  void feedback_callback(
-    GoalHandleFibonacci::SharedPtr,
-    const std::shared_ptr<const Fibonacci::Feedback> feedback)
-  {
-
-  }
-
-  void result_callback(const GoalHandleFibonacci::WrappedResult & result)
-  {
-
-  }
-};  // class FibonacciActionClient
-
-}  // namespace action_tutorials_cpp
-
-RCLCPP_COMPONENTS_REGISTER_NODE(action_tutorials_cpp::FibonacciActionClient)
