+++ source/Tutorials/Actions/client_0.py
@@ -11,11 +11,23 @@ class FibonacciActionClient(Node):
         super().__init__('fibonacci_action_client')
         self._action_client = ActionClient(self, Fibonacci, 'fibonacci')
 
+    def send_goal(self, order):
+        goal_msg = Fibonacci.Goal()
+        goal_msg.order = order
+
+        self._action_client.wait_for_server()
+
+        return self._action_client.send_goal_async(goal_msg)
+
 
 def main(args=None):
     rclpy.init(args=args)
 
     action_client = FibonacciActionClient()
+
+    future = action_client.send_goal(10)
+
+    rclpy.spin_until_future_complete(action_client, future)
 
 
 if __name__ == '__main__':
