--- source/Tutorials/Actions/client_1.c++
@@ -1,80 +1,0 @@
-#include <memory>
-#include <string>
-#include <sstream>
-
-#include "action_tutorials_interfaces/action/fibonacci.hpp"
-
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_action/rclcpp_action.hpp"
-#include "rclcpp_components/register_node_macro.hpp"
-
-namespace action_tutorials_cpp
-{
-class FibonacciActionClient : public rclcpp::Node
-{
-public:
-  explicit FibonacciActionClient(const rclcpp::NodeOptions & options)
-  : Node("fibonacci_action_client", options)
-  {
-    this->client_ptr_ = rclcpp_action::create_client<Fibonacci>(
-      this->get_node_base_interface(),
-      this->get_node_graph_interface(),
-      this->get_node_logging_interface(),
-      this->get_node_waitables_interface(),
-      "fibonacci");
-
-    this->timer_ = this->create_wall_timer(
-      std::chrono::milliseconds(500),
-      std::bind(&FibonacciActionClient::send_goal, this));
-  }
-
-  void send_goal()
-  {
-    using namespace std::placeholders;
-
-    this->timer_->cancel();
-
-    if (!this->client_ptr_->wait_for_action_server(std::chrono::seconds(10))) {
-      RCLCPP_ERROR(this->get_logger(), "Action server not available after waiting");
-      rclcpp::shutdown();
-    }
-
-    auto goal_msg = Fibonacci::Goal();
-    goal_msg.order = 10;
-
-    RCLCPP_INFO(this->get_logger(), "Sending goal");
-
-    auto send_goal_options = rclcpp_action::Client<Fibonacci>::SendGoalOptions();
-    send_goal_options.goal_response_callback =
-      std::bind(&FibonacciActionClient::goal_response_callback, this, _1);
-    send_goal_options.feedback_callback =
-      std::bind(&FibonacciActionClient::feedback_callback, this, _1, _2);
-    send_goal_options.result_callback =
-      std::bind(&FibonacciActionClient::result_callback, this, _1);
-    this->client_ptr_->async_send_goal(goal_msg, send_goal_options);
-  }
-private:
-  rclcpp_action::Client<Fibonacci>::SharedPtr client_ptr_;
-  rclcpp::TimerBase::SharedPtr timer_;
-
-  void goal_response_callback(std::shared_future<GoalHandleFibonacci::SharedPtr> future)
-  {
-
-  }
-
-  void feedback_callback(
-    GoalHandleFibonacci::SharedPtr,
-    const std::shared_ptr<const Fibonacci::Feedback> feedback)
-  {
-
-  }
-
-  void result_callback(const GoalHandleFibonacci::WrappedResult & result)
-  {
-
-  }
-};  // class FibonacciActionClient
-
-}  // namespace action_tutorials_cpp
-
-RCLCPP_COMPONENTS_REGISTER_NODE(action_tutorials_cpp::FibonacciActionClient)
