+++ source/Tutorials/Actions/client_1.py
@@ -17,11 +17,25 @@ class FibonacciActionClient(Node):
 
         self._action_client.wait_for_server()
 
-        self._action_client.send_goal_async(goal_msg)
+        self._send_goal_future = self._action_client.send_goal_async(goal_msg)
 
-    def feedback_callback(self, feedback_msg):
-        feedback = feedback_msg.feedback
-        self.get_logger().info('Received feedback: {0}'.format(feedback.partial_sequence))
+        self._send_goal_future.add_done_callback(self.goal_response_callback)
+
+    def goal_response_callback(self, future):
+        goal_handle = future.result()
+        if not goal_handle.accepted:
+            self.get_logger().info('Goal reject :(')
+            return
+
+        self.get_logger().info('Goal accepted :)')
+
+        self._get_result_future = goal_handle.get_result_async()
+        self._get_result_future.add_done_callback(self.get_result_callback)
+
+    def get_result_callback(self, future):
+        result = future.result().result
+        self.get_logger().info('Result: {0}'.format(result.sequence))
+        rclpy.shutdown()
 
 
 def main(args=None):
@@ -30,6 +44,8 @@ def main(args=None):
     action_client = FibonacciActionClient()
 
     action_client.send_goal(10)
+
+    rclpy.spin(action_client)
 
 
 if __name__ == '__main__':
