+++ source/Tutorials/Actions/server_0.py
@@ -17,7 +17,8 @@ class FibonacciActionServer(Node):
 
     def execute_callback(self, goal_handle):
         self.get_logger().info('Executing goal...')
-        return Fibonacci.Result()
+        result = Fibonacci.Result()
+        return result
 
 
 def main(args=None):
